changed name of common type
