+++ schema/schema-ruby-erb/Model Generation/Models/Accessibility/accessibility.rb
@@ -14,7 +14,7 @@
 
 require_relative '../../Synthax/Attributes/variable.rb'
 require_relative '../base_component.rb'
-require_relative '../../Synthax/Types/common_type.rb'
+require_relative '../../Synthax/Types/built_in_type.rb'
 
 class Accessibility < BaseComponent
 
@@ -24,7 +24,7 @@ class Accessibility < BaseComponent
             Variable.new(:name => "accessible", :typeName => "Bool", :defaultValue => "true")
 
         ]
-        synthaxType = CommonType.new(
+        synthaxType = BuiltInType.new(
             :kind => 'struct',
             :name => self.name,
             :variables => variables,
