+++ schema/schema-ruby-erb/Model Generation/Models/Layout/corner_radius.rb
@@ -14,7 +14,7 @@
 
 require_relative '../../Synthax/Attributes/variable.rb'
 require_relative '../base_component.rb'
-require_relative '../../Synthax/Types/common_type.rb'
+require_relative '../../Synthax/Types/built_in_type.rb'
 
 class CornerRadius < BaseComponent
 
@@ -22,7 +22,7 @@ class CornerRadius < BaseComponent
         textVariables = [
             Variable.new(:name => "radius", :typeName => "Double")
         ]
-        synthaxType = CommonType.new(
+        synthaxType = BuiltInType.new(
             :kind => 'struct',
             :name => self.name,
             :variables => textVariables,
