+++ schema/schema-ruby-erb/Model Generation/Models/Layout/display.rb
@@ -18,7 +18,6 @@ require_relative '../../Synthax/Types/enum_type.rb'
 
 class Display < BaseComponent
 
-    # TODO add support for enums to TS, Kotlin and Kotlin Backend
     def initialize
         displayCases = [
             EnumCase.new(:name => "flex", :defaultValue => "FLEX"),
