+++ schema/schema-ruby-erb/Model Generation/Models/Layout/size.rb
@@ -14,7 +14,7 @@
 
 require_relative '../../Synthax/Attributes/variable.rb'
 require_relative '../base_component.rb'
-require_relative '../../Synthax/Types/common_type.rb'
+require_relative '../../Synthax/Types/built_in_type.rb'
 
 class Size < BaseComponent
 
@@ -28,7 +28,7 @@ class Size < BaseComponent
             Variable.new(:name => "minHeight", :typeName => unitValue, :isOptional => true),
             Variable.new(:name => "aspectRatio", :typeName => "Double", :isOptional => true)
         ]
-        synthaxType = CommonType.new(
+        synthaxType = BuiltInType.new(
             :kind => 'struct',
             :name => self.name,
             :variables => textVariables,
