+++ schema/schema-ruby-erb/Model Generation/Models/Widgets/text.rb
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-require_relative '../../Synthax/Types/common_type.rb'
+require_relative '../../Synthax/Types/built_in_type.rb'
 require_relative '../../Synthax/Attributes/variable.rb'
 require_relative '../base_component.rb'
 require_relative './widget.rb'
@@ -24,7 +24,7 @@ class Text < BaseComponent
             Variable.new(:name => "text", :typeName => "String", :isBindable => true),
             Variable.new(:name => "styleId", :typeName => "String", :isOptional => true)
         ]
-        synthaxType = CommonType.new(
+        synthaxType = BuiltInType.new(
             :kind => 'struct',
             :name => self.name,
             :variables => textVariables,
