+++ schema/schema-ruby-erb/Model Generation/Synthax/Types/built_in_type.rb
@@ -12,42 +12,12 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-#todo WIP
-class SupportedLanguages
+require_relative './type.rb'
 
-    attr_accessor :swift, :kotlin, :kotlinBackend, :typeScript
-
-    @swift = "swift"
-    @kotlin = "kt"
-    @kotlinBackend = "ktBackend"
-    @typeScript = "ts"
-    
-    @dart = "dart" #todo 
-
-end
-
-class PrimitiveType
+class BuiltInType
     include BaseType
-    @sl = SupportedLanguages.new
-    
-    @string = [
-        @sl.swift => "String",
-        @sl.kotlin => "String"
-    ]
-
-    @bool = [
-        @sl.swift => "Bool",
-        @sl.kotlin => "Bool"
-    ]
-
-    @interface = [
-        @sl.swift => "protocol",
-        @sl.kotlin => "interface"
-    ]
-
-    @enum = [
-        @sl.swift => "enum",
-        @sl.kotlin => "enum"
-    ]
+    def initialize(params = {})
+        super
+    end
 
 end
